requirements.txt
PyYAML
mysql-connector-python
boto3
